# backend/doc/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: ProdCom-E API
  description: An e-commerce website to deliver best and authentic gadget items.
  contact:
    name: Md. Sifytul Karim
    email: sifytul09@gmail.com
...
servers:
  - description: localhost
    url: http://localhost:4000
paths:
...
  '/products': &all_products_endpoint
    get:
      description: This route will retrieve all the products
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/categoryParam'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/sordTypeParam'
        - $ref: '#/components/parameters/sortByParam'
        - $ref: '#/components/parameters/searchParam'
      responses:
        '200':
          description: retrieve all articles based on the query param
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        sku:
                          type: string
                          example: abccd-11
                        name:
                          type: string
                          example: Macbook Pro
                        image_url:
                          type: string
                          example: s3.productimage.jpeg
                        price:
                          type: number
                          example: 499.99
                        ratings:
                          type: number
                          example: 4.8
                        stock:
                          type: integer
                          example: 20
                        links:
                          type: object
                          properties:
                            self:
                              type: string
                              example: /products/2
        '500':
          $ref: '#/components/responses/500'
  '/products/{id}':
    get:
      description: Retrieve single product by id
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/productIdParam'
      responses:
        '200':
          description: success response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      sku:
                        type: string
                        example: abccd-11
                      name:
                        type: string
                        example: Macbook Pro
                      description:
                        type: string
                        example: This is an awesome product
                      variant:
                        type: string
                        example: China
                      colors:
                        type: array
                        example:
                          - blue
                          - white
                      sizes:
                        type: string
                        example:
                      image_urls:
                        type: array
                        example:
                          - s3.productimage.jpeg
                          - s3.secondimage.jpeg
                      price:
                        type: number
                        example: 499.99
                      ratings:
                        type: number
                        example: 4.8
                      stock:
                        type: integer
                        example: 20
                      reviews:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Review'
                            - $ref: '#/components/schemas/Timestamp'
        '404':
          $ref: '#/components/responses/404'

  # Products - Admin Endpoints
...
  '/admin/products/new':
    post:
      description: This route will create a product with provided payload.
      tags:
        - Products
      security:
        - bearerAuth: []
      requestBody:
        description: product payload
        content:
          'application/json':
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: xiaomi airdot 6
                description:
                  type: string
                  example: Details description of the product
                price:
                  type: number
                  example: 29.99
                category:
                  type: string
                  example: phone
                stock:
                  type: integer
                  example: 30
              required:
                - name
                - description
                - price
                - image_urls
                - category
          'multipart/form-data':
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: xiaomi airdot 6
                description:
                  type: string
                  example: Details description of the product
                price:
                  type: number
                  example: 29.99
                images:
                  description: only jpg, jpeg, png and file less than 1mb is allowed
                  type: array
                  items:
                    type: string
                    format: binary
                category:
                  type: string
                  example: phone
                stock:
                  type: integer
                  example: 30
              required:
                - name
                - description
                - price
                - image_urls
                - category
            encoding:
              images:
                contentType: image/jpeg, image/png, image/jpg
      responses:
        '201':
          description: success response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: successfully added
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  '/admin/products/{id}':
    patch:
      description: This route will update a product based on productId
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/productIdParam'
      requestBody:
        description: product update payload
        content:
          'application/json':
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: xiaomi airdot 6
                description:
                  type: string
                  example: Details description of the product
                price:
                  type: number
                  example: 29.99
                image_urls:
                  type: array
                  example:
                    - s3.firstimage.jpeg
                    - s3.secondimage.jpeg
                category:
                  type: string
                  example: phone
                stock:
                  type: integer
                  example: 30
      responses:
        '200':
          $ref: '#/components/responses/200'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
    delete:
      description: This route will remove temporarily a product based on productId
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/productIdParam'
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  '/admin/products/{id}/hard-delete':
    delete:
      description: This route will permanently delete a product based on productId
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/productIdParam'
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  '/admin/products/soft-deleted':
    get:
      description: This route will retrieve all the soft deleted products.
      tags:
        - Products
      security:
        - bearerAuth: []
      responses:
        '200':
          description: retrieve all soft deleted products
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        sku:
                          type: string
                          example: abccd-11
                        name:
                          type: string
                          example: Macbook Pro
                        description:
                          type: string
                          example: This is an awesome product
                        variant:
                          type: string
                          example: China
                        image_urls:
                          type: array
                          example:
                            - s3.productimage.jpeg
                            - s3.secondimage.jpeg
                        price:
                          type: number
                          example: 499.99
                        ratings:
                          type: number
                          example: 4.8
                        stock:
                          type: integer
                          example: 20
                        category:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            category_name:
                              type: string
                              example: phone
                            description:
                              type: string
                              example: This is a phone category.
                            image_url:
                              type: string
                              example: s3.categoryimage.jpeg
                        deleted_at:
                          type: string
                          example: 2023-10-13T09:25:14.836Z
                        created_at:
                          type: string
                          example: 2023-10-13T09:25:14.836Z
                        updated_at:
                          type: string
                          example: 2023-10-13T09:25:14.836Z
                        links:
                          type: object
                          properties:
                            self:
                              type: string
                              example: /products/2
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  '/admin/products/{id}/restore':
    patch:
      description: This route will restore a soft deleted product based on productId.
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/productIdParam'
      responses:
        '200':
          description: success response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: successfully restored
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'

  # Category
...
components:
  responses:
...
    '401':
      description: Unauthorized
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Please sign in to perform the action.
    '403':
      description: Forbidden
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: You are not authorized to see these resources
    '500':
      description: Server error
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Internal server Error
    '404':
      description: Resource not found
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: not found
    '204':
      description: successful deletion of the resource
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: Successfully performed the delete action
    '200':
      description: update success response
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: successfully updated
  parameters:
...
    categoryParam:
      in: query
      name: category
      required: false
      description: category
      schema:
        type: string
        example: gadgets
    pageParam:
      in: query
      name: page
      required: false
      description: page no
      schema:
        type: integer
        format: int32
        example: 2
    limitParam:
      in: query
      name: limit
      required: false
      description: number of users in one page
      schema:
        type: integer
        format: int32
        example: 10
    sordTypeParam:
      in: query
      name: sort_type
      description: how the data will be sorted (asc or desc)
      schema:
        type: string
        enum: [asc, desc]
        example: desc
    sortByParam:
      in: query
      name: sort_by
      description: by which field the users will be sorted(by createdAt or name)
      schema:
        type: string
        enum: [created_at, name]
        example: created_at
    searchParam:
      in: query
      name: search
      description: search term
      schema:
        type: string
        example: sifytul
    productIdParam:
      in: path
      name: id
      required: true
      description: product ID
      schema:
        type: integer
        format: int32
        example: 2
...
  schemas:
...
    Review:
      type: object
      properties:
        reviewerId:
          type: integer
          example: 5
        reviewerName:
          type: string
          example: Khan
        comment:
          type: string
          example: Nice Product
        rating:
          type: integer
          example: 5
          minimum: 0
          maximum: 5
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          example: '2023-9-3 03:03:00'
        updatedAt:
          type: string
          example: '2023-9-3 03:03:00'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT